refactor: rename package 'service' to 'filterService'
